Moved the code to electron file but don't work
